+++ proyecto-limpio-spring/src/main/java/ar/edu/unlam/tallerweb1/servicios/Interfaces/IServicioTarjeta.java
@@ -4,5 +4,5 @@ import ar.edu.unlam.tallerweb1.modelo.Tarjeta;
 
 public interface IServicioTarjeta {
 
-    Tarjeta consultarTarjeta(int numeroTarjeta);
+    Tarjeta consultarTarjeta(Long numeroTarjeta);
 }
